requirements.txt
hvac>=2.1.0
hydra-core>=1.3.2
numpy<2,>=1.22.4
pandas
azure-keyvault-secrets>=4.8.0
azure-identity>=1.15.0
PyGithub>=2.2.0
PyGithub>=2.3.0
cryptography>=42.0.5
azure-mgmt-resource>=23.0.1
msticpy>=2.10.0
azure.mgmt.network>=25.3.0
azure-mgmt-monitor>=6.0.2
azure-mgmt-compute>=30.6.0
boto3>=1.34.74
OPA-python-client>=1.3.2
pytest>=6.2.0
pytest-cov>=3.0.0
sphinx>=7.2.6
sphinx-autoapi>=3.0.0
sphinx-rtd-theme>=2.0.0
azure-mgmt-authorization>=4.0.0
azure-mgmt-subscription>=3.1.1
ansible>=9.4.0
ansible-runner>=2.3.6
ansible-base>=2.10.17 
ansible-lint>=24.2.2